refactor: split parser modules
