+++ crates/ox_content_parser/src/parser/inline/scan.rs
@@ -1,0 +1,49 @@
+/// Lookup table: `INLINE_SPECIAL[b] == 1` iff the byte can begin an inline
+/// construct handled by `parse_inline_special`.
+///
+/// The table deliberately stores flags instead of enum variants. The scan in
+/// `next_inline_special` ORs eight table entries at a time, which gives LLVM a
+/// simple branch-free loop for long runs of normal text. The actual parser
+/// decision still happens only after a candidate byte is found.
+static INLINE_SPECIAL: [u8; 256] = {
+    let mut t = [0u8; 256];
+    t[b'*' as usize] = 1;
+    t[b'_' as usize] = 1;
+    t[b'`' as usize] = 1;
+    t[b'[' as usize] = 1;
+    t[b'!' as usize] = 1;
+    t[b'~' as usize] = 1;
+    t[b'\\' as usize] = 1;
+    t[b'<' as usize] = 1;
+    t
+};
+
+#[inline]
+pub(super) fn next_inline_special(bytes: &[u8], from: usize) -> usize {
+    // Skip eight bytes at a time while the OR of their marker flags is zero.
+    // This is not a semantic parser: it only proves that none of those bytes
+    // can start inline syntax, so returning the first flagged byte preserves
+    // the exact same marker positions as the previous per-byte loop.
+    let mut i = from;
+    let end = bytes.len();
+
+    while i + 8 <= end {
+        let chunk = &bytes[i..i + 8];
+        let mask = INLINE_SPECIAL[chunk[0] as usize]
+            | INLINE_SPECIAL[chunk[1] as usize]
+            | INLINE_SPECIAL[chunk[2] as usize]
+            | INLINE_SPECIAL[chunk[3] as usize]
+            | INLINE_SPECIAL[chunk[4] as usize]
+            | INLINE_SPECIAL[chunk[5] as usize]
+            | INLINE_SPECIAL[chunk[6] as usize]
+            | INLINE_SPECIAL[chunk[7] as usize];
+        if mask != 0 {
+            break;
+        }
+        i += 8;
+    }
+    while i < end && INLINE_SPECIAL[bytes[i] as usize] == 0 {
+        i += 1;
+    }
+    i
+}
